Refactor CLI organisation, and improve termination
